pyproject.toml
[tool.poetry]
name = "aiogram-bot-template"
version = "0.1.0"
description = ""
authors = ["rkjar <rkjar@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"
redis = "^6.4.0"
httpx = "^0.28.1"
pydantic = "^2.11.7"


[tool.poetry.group.dev.dependencies]
setuptools = "^80.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
